face_attack/requirements.txt
opencv-python==4.1.2.30
Pillow==6.2.1
face-recognition==1.2.3
numba==0.46.0
scipy==1.2.1
joblib==0.14.0